Fix VCL tempalte replacement for querystring.regfilter()
If users do not want to strip any query parameters and decided to leave the UI field as blank (remove all default values), that will produce "querystring.regfilter(req.url, "^()");" which matches any query parameters and eventually strip all parameters. To remediate this, we set "&" in case if the field is blank so this doesn't match any query parameters as "&" is a delimiter of query strings.